add HeroesService
