+++ package.json
@@ -14,6 +14,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-scripts": "5.0.1",
+    "redux-thunk": "^2.4.2",
     "reselect": "^4.1.8",
     "sass": "^1.62.1",
     "uuid": "^9.0.0",
