+++ src/components/heroesFilters/HeroesFilters.js
@@ -7,21 +7,19 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from "react-redux";
 import classNames from 'classnames';
 
-import { activeFilterChanged, filtersFetched, filtersFetching, filtersFetchingError } from '../../actions';
-import { useHttp } from "../../hooks/http.hook";
+import { activeFilterChanged } from '../../actions';
 import { Spinner } from '../spinner/Spinner';
+import { useHeroesService } from '../../services/HeroesService';
 
 export const HeroesFilters = () => {
 
   const { filters, filtersLoadingStatus, activeFilter } = useSelector((state) => state.filtersReducer);
   const dispatch = useDispatch();
-  const { request } = useHttp();
+  const { getFilters } = useHeroesService();
 
   useEffect(() => {
-    dispatch(filtersFetching());
-    request('http://localhost:3001/filters')
-      .then((data) => dispatch(filtersFetched(data)))
-      .catch(() => dispatch(filtersFetchingError()));
+    getFilters();
+    // eslint-disable-next-line
   }, []);
 
   if (filtersLoadingStatus === "loading") {
