+++ src/hooks/http.hook.js
@@ -3,12 +3,12 @@ import { useCallback } from "react";
 export const useHttp = () => {
     // const [process, setProcess] = useState('waiting');
 
-    const request = useCallback(async (url, method = 'GET', body = null, headers = {'Content-Type': 'application/json'}) => {
+    const request = useCallback(async (url, method = 'GET', body = null, headers = { 'Content-Type': 'application/json' }) => {
 
         // setProcess('loading');
 
         try {
-            const response = await fetch(url, {method, body, headers});
+            const response = await fetch(url, { method, body, headers });
 
             if (!response.ok) {
                 throw new Error(`Could not fetch ${url}, status: ${response.status}`);
@@ -17,19 +17,20 @@ export const useHttp = () => {
             const data = await response.json();
 
             return data;
-        } catch(e) {
+        } catch (e) {
             // setProcess('error');
             throw e;
         }
     }, []);
 
     // const clearError = useCallback(() => {
-        // setProcess('loading');
+    // setProcess('loading');
     // }, []);
 
-    return {request, 
-            // clearError, 
-            // process, 
-            // setProcess
-        }
+    return {
+        request,
+        // clearError, 
+        // process, 
+        // setProcess
+    }
 }
